Basic working set up for add/view/update goals

+++ src/app/services/resolution-service.ts
@@ -1,4 +1,5 @@
 import Resolution, { SerializedResolution } from "src/app/models/resolution";
+import { ProgressRecord } from "../models/progress-record";
 
 const RESOLUTIONS_KEY = "RESOLUTIONS";
 
@@ -23,6 +24,28 @@ class ResolutionService {
             const existingResolutions =
                 jsonString != null ? JSON.parse(jsonString) : [];
             existingResolutions.push(resolution.serialize());
+            localStorage.setItem(
+                RESOLUTIONS_KEY,
+                JSON.stringify(existingResolutions)
+            );
+        } catch (error) {
+            console.warn(error);
+        }
+    };
+
+    static addRecord = (id: string, record: ProgressRecord) => {
+        try {
+            const jsonString = localStorage.getItem(RESOLUTIONS_KEY);
+            const existingResolutions =
+                jsonString != null ? JSON.parse(jsonString) : [];
+            const existingIndex = existingResolutions.findIndex(
+                (e: SerializedResolution) => e.id === id
+            );
+            if (existingIndex === -1) {
+                console.warn("No such element");
+                return;
+            }
+            existingResolutions[existingIndex].records.push(record);
             localStorage.setItem(
                 RESOLUTIONS_KEY,
                 JSON.stringify(existingResolutions)
